refactor: 💡 globale email validation helper
