+++ src/components/message/FurtherSteps.tsx
@@ -15,10 +15,7 @@ import {
 	InputFieldItem,
 	InputFieldLabelState
 } from '../inputField/InputField';
-import {
-	isStringValidEmail,
-	ResortData
-} from '../registration/registrationHelpers';
+import { ResortData } from '../registration/registrationHelpers';
 import { useContext, useEffect, useState } from 'react';
 import {
 	Overlay,
@@ -26,7 +23,7 @@ import {
 	OverlayWrapper,
 	OVERLAY_FUNCTIONS
 } from '../overlay/Overlay';
-import { apiPutEmail, FETCH_ERRORS } from '../../api';
+import { apiPutEmail, FETCH_ERRORS, validateEmail } from '../../api';
 import {
 	ActiveSessionGroupIdContext,
 	getActiveSession,
@@ -86,27 +83,6 @@ export const FurtherSteps = (props: FurtherStepsProps) => {
 		name: 'email',
 		type: 'text',
 		labelState: emailLabelState
-	};
-
-	const validateEmail = (
-		email
-	): { valid: InputFieldLabelState; label: string } => {
-		if (email.length > 0 && isStringValidEmail(email)) {
-			return {
-				valid: 'valid',
-				label: translate('furtherSteps.email.overlay.input.valid')
-			};
-		} else if (email.length > 0) {
-			return {
-				valid: 'invalid',
-				label: translate('furtherSteps.email.overlay.input.invalid')
-			};
-		} else {
-			return {
-				valid: null,
-				label: translate('furtherSteps.email.overlay.input.label')
-			};
-		}
 	};
 
 	const handleEmailChange = (event) => {
